RavenDB-19071 support for serverSideQuery
